+++ src/App.vue
@@ -33,9 +33,19 @@
       />
     </div>
 
+    <div class="group">
+      <im-button
+        type="warn"
+        @click="testConfirm"
+      >
+        testConfirm
+      </im-button>
+    </div>
+
     <div class="group" style="margin-top: 100px;">
       <im-slider v-model="sliderValue"></im-slider>
     </div>
+    
   </div>
 </template>
 
@@ -64,6 +74,14 @@ export default {
       this.$Message({
         message: 'test',
         title: 'test'
+      });
+    },
+    testConfirm() {
+      this.$Confirm({
+        message: 'test',
+        title: 'test',
+        onConfirm: () => { console.log('confirm') },
+        onCancel: () => { console.log('cancel') }
       });
     },
     handleSwitchChange() {